refactor: waiting-room 내부 파일 구조 정리
