+++ src/pages/waiting-room/page/WaitingRoomFlow.test.tsx
@@ -2,13 +2,13 @@ import { screen, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { Route, Routes, useParams } from 'react-router-dom'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import { useAuthStore } from '../../features/auth/store'
-import { createAuthSessionFixture } from '../../test/fixtures'
-import { renderWithProviders } from '../../test/renderWithProviders'
-import LobbyPage from '../lobby/LobbyPage'
+import { useAuthStore } from '../../../features/auth/store'
+import { createAuthSessionFixture } from '../../../test/fixtures'
+import { renderWithProviders } from '../../../test/renderWithProviders'
+import LobbyPage from '../../lobby/LobbyPage'
 import WaitingRoomPage from './WaitingRoomPage'
 
-vi.mock('../../config/env', () => ({
+vi.mock('../../../config/env', () => ({
   IS_SOCKET_MOCK_ENABLED: true,
 }))
 
@@ -22,16 +22,16 @@ const {
   useDirectMessageControllerMock: vi.fn(),
 }))
 
-vi.mock('../lobby/hooks', () => ({
+vi.mock('../../lobby/hooks', () => ({
   useLobbyRoomsQuery: useLobbyRoomsQueryMock,
 }))
 
-vi.mock('../../features/presence/online-users/useOnlineUsersSocket', () => ({
+vi.mock('../../../features/presence/online-users/useOnlineUsersSocket', () => ({
   useOnlineUsersSocket: useOnlineUsersSocketMock,
 }))
 
 vi.mock(
-  '../../features/presence/direct-message/useDirectMessageController',
+  '../../../features/presence/direct-message/useDirectMessageController',
   () => ({
     useDirectMessageController: useDirectMessageControllerMock,
   })
